+++ src/core/object/watch/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.62.1 (2021-11-08)
+
+#### :bug: Bug Fix
+
+* Fixed watching for an array by a complex path with indices
+
 ## v3.56.1 (2021-07-28)
 
 #### :bug: Bug Fix